Refactor all the things
- Completely refactor StoryMaster
- Make viewing and joining unjoined frequencies possible (still requires
a page refresh-for now)
- Refactor the way we fetch stories and frequencies

Previously we only fetched a users frequencies and the associated
stories, but the issue is that means we can't show unjoined
frequencies, as the data won't be there.

This PR makes it so that all frequencies and stories are fetched and
cached. We need to change this to fetch frequencies on the fly when they
are opened, but that can happen down the line for now.

This generally refactors a bunch of stuff and improves the data flow
throughout the app.

Closes #183 (no longer relevant, this is working fine now), closes #179,
closes #130