Support link previews in composer, story cards, and stories.
When a user types a URL into the composer, we fetch OG meta tags from that website and return what we can in order to populate a link preview. That data gets sent to the composer stor,e and then when published it goes to the db as a metadata field on the story object.

Currently doesn't do a great job of handling failed url checks, but I will work on this going forward.