Fix mac notifications (#1252)
* Fix mac notifications

* Fix push-sw not updating

It turns out serviceworker additions added with `importScripts` (like our push serviceworker) only get updated when the serviceworker gets updated. Since our ServiceWorker is automatically generated it doesn't actually get updated all that often, meaning we have to do it manually.

The way to update an `importScripts`ed file is to change it's filename, so that's the way to change PushSW now. The issue is that you'd also have to update the build script everytime, so I work around that by using a glob to get the file matching `push-sw-*.js` so you only have to change the filename for it to update.