Don't mark notifications sent via push as seen
This behaviour has unfortunately turned out to be a bit of a PITA. I'll
go visit Spectrum on my laptop, not looking at my phone, and I'll be
missing notifications because I got them as web push but I hadn't looked
at them.

This is just as fine and will be a much better experience! :+1: