Maybe fix message flicker
In our optimistic response we were settings `isOwner`, `isModerator`,
`reputation` etc just to default values, but that meant when the
opmistic response was displayed it would show the user with 0 reputation
across the whole app.

By just getting rid of that part we display the values that Apollo has
cached locally, which are the right ones anyway.

This might also fix message flickering.