Partially fix linkification
Now both markdown and normal links are linkified as they should when you
write threads.

The last remaining bug is that markdown links still aren't highlighted
when you are writing a thread, that is a decorator issue and will be
fixed with the upcoming refactor of the editor component. (same issue as
code highlighting not working when you're typing)