Fix DraftJS on Android
Somebody submitted 5 line patch upstream to DraftJS which (should) fix
DraftJS on Android. (https://github.com/facebook/draft-js/pull/1500)

I tested this locally and it works super well on my phone, so I pulled
the repo locally and deployed a quick fork. This PR just sets the
package `draft-js` to install the fork based on that PR instead.

Note that this isn't any kind of long-term solution, but it finally
fixes the problem and let's me respond to messages from my phone, so I
say fuck long-term and let's ship this.