Store all session data in cookie
Switches us from storing all session data in RethinkDB to just storing
it all in the cookie. This means we store way less data in RethinkDB and
query it a bunch of times less :tada:

This also still works the exact same as before. Oh, and logging out
works now.

Boom shakalaka babeh