Channel changes must also update indexed messages
I realized that because we are indexing messages and threads together in the same place on Algolia, we need to account for more scenarios than I'd originally thought. This commit makes it so that if you change a thread's channel, it will also update all of the indexed messages on that thread to reflect the same new channel id in Algolia. Obviously we need this so that if you move a thread to a private channel, those messages won't trigger thread results in public thread searches