Fix scrolling on thread message pagination
We didn't find this bug since the custom scrolling of thread message was
completely fucked due to a bug I fixed in the optimstic update branch,
but anytime you'd load in any messages either before or after the
current messages it would scroll you to the bottom.

That feels very bad. (try paginating messages on live. It's horrible.)

This fixes it by specifically only scrolling you to the bottom on first
load of the messages (if you're a participant) or if a single new
message comes in. (i.e. if you or another person sends one, so the chat
input "sticks" to the bottom)