Start Apollo engine after webserver
Zeit ends the deployment process as soon as a port is opened. Apollo
Engine opens a port to proxy the requests through to the API, but really
we don't want Zeit to finish the deployment process until the whole
server is instantiated and the API is running.

This patch moves Apollo Engine starting to happen after the API server
is already running.